# frontend/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CampusEats v1",
    "description": "单校区点餐演示；所有支付退款均为模拟",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/user/auth/wechat": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "wx",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WxLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SessionView"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/auth/demo": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "demo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SessionView"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/employees": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "employees",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EmployeeView"
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "employee",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Employee"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/auth/password": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Password"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/user/auth/logout": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/auth/logout": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "logout_1",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SessionView"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/user/auth/me": {
      "get": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Actor"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/auth/me": {
      "get": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "me_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Actor"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Employee": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 32,
            "minLength": 0
          },
          "name": {
            "type": "string",
            "maxLength": 32,
            "minLength": 0
          },
          "role": {
            "type": "string",
            "minLength": 1,
            "pattern": "ADMIN|OPERATOR|DELIVERER"
          },
          "enabled": {
            "type": "boolean"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12
          }
        },
        "required": [
          "name",
          "role",
          "username"
        ]
      },
...
      "WxLogin": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 256,
            "minLength": 0
          }
        },
        "required": [
          "code"
        ]
      },
      "SessionView": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "OPERATOR",
              "DELIVERER",
              "USER"
            ]
          },
          "mustChangePassword": {
            "type": "boolean"
          },
          "expiresAt": {
            "type": "string"
          }
        }
      },
      "DemoLogin": {
        "type": "object",
        "properties": {
          "account": {
            "type": "integer",
            "format": "int32",
            "maximum": 3,
            "minimum": 1
          }
        }
      },
      "Password": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "maxLength": 128,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12
          }
        },
        "required": [
          "oldPassword",
          "password"
        ]
      },
      "Login": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 32,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 0
          }
        },
        "required": [
          "password",
          "username"
        ]
      },
...
      "EmployeeView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "mustChangePassword": {
            "type": "boolean"
          }
        }
      },
      "Actor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "type": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "jti": {
            "type": "string"
          },
          "mustChangePassword": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}